requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
branca==0.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.1.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
filelock==3.16.1
flatbuffers==24.3.25
folium==0.19.1
frozendict==2.4.6
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencage==3.0.3
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
phonenumbers==8.13.52
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.1
pycparser==2.22
PyMatting==1.1.13
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
ratelim==0.1.6
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.22.3
schedule==1.2.2
scikit-image==0.24.0
scipy==1.14.1
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.18.3
yfinance==0.2.50
